Implemented AddButton component
